@@ -13,39 +13,40 @@
 	},
 	"devDependencies": {
 		"@capacitor/assets": "^3.0.5",
-		"@sveltejs/adapter-auto": "^3.2.4",
-		"@sveltejs/adapter-static": "^3.0.4",
-		"@sveltejs/kit": "^2.5.26",
-		"@sveltejs/vite-plugin-svelte": "^3.1.2",
+		"@sveltejs/adapter-auto": "^3.3.1",
+		"@sveltejs/adapter-static": "^3.0.6",
+		"@sveltejs/kit": "^2.8.5",
+		"@sveltejs/vite-plugin-svelte": "^4.0.2",
 		"autoprefixer": "^10.4.20",
 		"copyfiles": "^2.4.1",
-		"prettier": "^3.3.3",
-		"prettier-plugin-svelte": "^3.2.6",
-		"prettier-plugin-tailwindcss": "^0.6.6",
+		"prettier": "^3.4.1",
+		"prettier-plugin-svelte": "^3.3.2",
+		"prettier-plugin-tailwindcss": "^0.6.9",
 		"rimraf": "^6.0.1",
-		"svelte": "^4.2.19",
-		"svelte-check": "^3.8.6",
-		"tailwindcss": "^3.4.10",
-		"typescript": "^5.5.4",
-		"vite": "^5.4.3"
+		"svelte": "^5.2.10",
+		"svelte-check": "^4.1.0",
+		"tailwindcss": "^3.4.15",
+		"tailwindcss-animate": "^1.0.7",
+		"typescript": "^5.7.2",
+		"vite": "^5.4.11"
 	},
 	"type": "module",
 	"dependencies": {
-		"@capacitor/android": "^6.1.2",
-		"@capacitor/app": "^6.0.1",
-		"@capacitor/cli": "^6.1.2",
-		"@capacitor/core": "^6.1.2",
-		"@internationalized/date": "^3.5.5",
-		"bits-ui": "^0.21.13",
+		"@capacitor/android": "^6.2.0",
+		"@capacitor/app": "^6.0.2",
+		"@capacitor/cli": "^6.2.0",
+		"@capacitor/core": "^6.2.0",
+		"@internationalized/date": "^3.6.0",
+		"bits-ui": "^1.0.0-next.64",
 		"clsx": "^2.1.1",
 		"date-fns": "^4.1.0",
-		"lucide-svelte": "^0.437.0",
-		"mode-watcher": "^0.4.1",
-		"svelte-i18n": "^4.0.0",
+		"lucide-svelte": "^0.462.0",
+		"mode-watcher": "^0.5.0",
+		"svelte-i18n": "^4.0.1",
 		"svelte-image-viewer": "^2.0.0",
 		"svelte-sonner": "^0.3.28",
-		"tailwind-merge": "^2.5.2",
-		"tailwind-variants": "^0.2.1",
-		"vaul-svelte": "^0.3.2"
+		"tailwind-merge": "^2.5.5",
+		"tailwind-variants": "^0.3.0",
+		"vaul-svelte": "^1.0.0-next.3"
 	}
 }
