@@ -1,18 +1,16 @@
 <script lang="ts">
-	import { Drawer as DrawerPrimitive } from 'vaul-svelte';
-	import { cn } from '$lib/utils.js';
+	import { Drawer as DrawerPrimitive } from "vaul-svelte";
+	import { cn } from "$lib/utils.js";
 
-	type $$Props = DrawerPrimitive.TitleProps;
-
-	export let el: $$Props['el'] = undefined;
-	let className: $$Props['class'] = undefined;
-	export { className as class };
+	let {
+		ref = $bindable(null),
+		class: className,
+		...restProps
+	}: DrawerPrimitive.TitleProps = $props();
 </script>
 
 <DrawerPrimitive.Title
-	bind:el
-	class={cn('text-lg font-semibold leading-none tracking-tight', className)}
-	{...$$restProps}
->
-	<slot />
-</DrawerPrimitive.Title>
+	bind:ref
+	class={cn("text-lg font-semibold leading-none tracking-tight", className)}
+	{...restProps}
+/>
