@@ -1,12 +1,12 @@
 <script lang="ts">
-	import { Drawer as DrawerPrimitive } from 'vaul-svelte';
+	import { Drawer as DrawerPrimitive } from "vaul-svelte";
 
-	type $$Props = DrawerPrimitive.Props;
-	export let shouldScaleBackground: $$Props['shouldScaleBackground'] = true;
-	export let open: $$Props['open'] = false;
-	export let activeSnapPoint: $$Props['activeSnapPoint'] = undefined;
+	let {
+		shouldScaleBackground = true,
+		open = $bindable(false),
+		activeSnapPoint = $bindable(null),
+		...restProps
+	}: DrawerPrimitive.RootProps = $props();
 </script>
 
-<DrawerPrimitive.Root {shouldScaleBackground} bind:open bind:activeSnapPoint {...$$restProps}>
-	<slot />
-</DrawerPrimitive.Root>
+<DrawerPrimitive.Root {shouldScaleBackground} bind:open bind:activeSnapPoint {...restProps} />
