@@ -1,26 +1,51 @@
+<script lang="ts" module>
+	import { type VariantProps, tv } from "tailwind-variants";
+
+	export const toggleVariants = tv({
+		base: "ring-offset-background hover:bg-muted hover:text-muted-foreground focus-visible:ring-ring data-[state=on]:bg-accent data-[state=on]:text-accent-foreground inline-flex items-center justify-center gap-2 rounded-md text-sm font-medium transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 [&_svg]:pointer-events-none [&_svg]:size-4 [&_svg]:shrink-0",
+		variants: {
+			variant: {
+				default: "bg-transparent",
+				outline:
+					"border-input hover:bg-accent hover:text-accent-foreground border bg-transparent",
+			},
+			size: {
+				default: "h-10 min-w-10 px-3",
+				sm: "h-9 min-w-9 px-2.5",
+				lg: "h-11 min-w-11 px-5",
+			},
+		},
+		defaultVariants: {
+			variant: "default",
+			size: "default",
+		},
+	});
+
+	export type ToggleVariant = VariantProps<typeof toggleVariants>["variant"];
+	export type ToggleSize = VariantProps<typeof toggleVariants>["size"];
+	export type ToggleVariants = VariantProps<typeof toggleVariants>;
+</script>
+
 <script lang="ts">
-	import { Toggle as TogglePrimitive } from 'bits-ui';
-	import { type Size, type Variant, toggleVariants } from './index.js';
-	import { cn } from '$lib/utils.js';
+	import { Toggle as TogglePrimitive } from "bits-ui";
+	import { cn } from "$lib/utils.js";
 
-	type $$Props = TogglePrimitive.Props & {
-		variant?: Variant;
-		size?: Size;
-	};
-	type $$Events = TogglePrimitive.Events;
-
-	let className: $$Props['class'] = undefined;
-	export let variant: $$Props['variant'] = 'default';
-	export let size: $$Props['size'] = 'default';
-	export let pressed: $$Props['pressed'] = undefined;
-	export { className as class };
+	let {
+		ref = $bindable(null),
+		pressed = $bindable(false),
+		class: className,
+		size = "default",
+		variant = "default",
+		...restProps
+	}: TogglePrimitive.RootProps & {
+		variant?: ToggleVariant;
+		size?: ToggleSize;
+	} = $props();
 </script>
 
 <TogglePrimitive.Root
+	bind:ref
 	bind:pressed
 	class={cn(toggleVariants({ variant, size, className }))}
-	{...$$restProps}
-	on:click
->
-	<slot />
-</TogglePrimitive.Root>
+	{...restProps}
+/>
