@@ -22,7 +22,9 @@
 	import { setMode } from 'mode-watcher';
 	const platform = Capacitor.getPlatform();
 
-	$: $currentTab = $t('titles.settings');
+	$effect(() => {
+		$currentTab = $t('titles.settings');
+	});
 
 	async function importContent() {
 		dataService.importContent();
@@ -37,10 +39,10 @@
 		database.exportToStorage();
 	}
 
-	function onLanguageSelectedChanged(selected: any) {
-		$locale = selected.value;
+	function onLanguageSelectedChanged(selected: string) {
+		$locale = selected;
 
-		Preferences.setString({ key: 'language', value: selected.value });
+		Preferences.setString({ key: 'language', value: selected });
 	}
 
 	async function onShowNotificationsChanged(checked: boolean) {
@@ -69,44 +71,54 @@
 		Preferences.setString({ key: 'notification-time', value: time });
 	}
 
-	async function onIntervalChanged(selected: any) {
+	async function onIntervalChanged(selected: string) {
+		const selectedNumber = parseInt(selected);
 		await dataService.setNotificationOptions({
 			allow: $showNotifications,
 			time: $notificationTime,
-			interval: selected.value
+			interval: selectedNumber
 		});
 
-		$notificationInterval = selected.value;
-		Preferences.setLong({ key: 'notification-interval', value: selected.value });
+		$notificationInterval = selectedNumber;
+		Preferences.setLong({ key: 'notification-interval', value: selectedNumber });
 	}
 
-	const intervalTimes = new Map<number, string>([
-		[86400000, $t('settings.notification-intervals.one')],
-		[43200000, $t('settings.notification-intervals.two')],
-		[21600000, $t('settings.notification-intervals.three')]
+	const intervalTimes = new Map<string, string>([
+		['86400000', $t('settings.notification-intervals.one')],
+		['43200000', $t('settings.notification-intervals.two')],
+		['21600000', $t('settings.notification-intervals.three')]
 	]);
 
-	function onThemeSelected(selected: any) {
-		$globalTheme = selected.value;
-		Preferences.setString({ key: 'theme', value: selected.value });
-		setMode(selected.value);
+	function onThemeSelected(selected: string) {
+		$globalTheme = selected;
+		Preferences.setString({ key: 'theme', value: selected });
+		setMode(selected as any);
 	}
+
+	let selectedLanguage = $state($locale);
+	const languageTriggerContent = $derived($t('settings.languages.' + selectedLanguage));
+
+	let selectedTheme = $state($globalTheme);
+	const themeTriggerContent = $derived($t('settings.themes.' + selectedTheme));
+
+	let selectedInterval = $state($notificationInterval.toString());
+	let intervalTriggerContent = $derived(intervalTimes.get(selectedInterval));
 
 	const themeSettings = ['system', 'dark', 'light'];
 </script>
 
 <div class="grid w-full gap-6 overflow-y-auto px-8 py-10">
 	<div class="grid gap-4 px-3">
-		<Button on:click={importContent} variant="outline" class="p-5 font-semibold"
+		<Button onclick={importContent} variant="outline" class="p-5 font-semibold"
 			><Import class="mr-3" size={20} />{$t('settings.import-content')}
 		</Button>
 
-		<Button on:click={exportContent} variant="outline" class="p-5"
+		<Button onclick={exportContent} variant="outline" class="p-5"
 			>{$t('settings.export-content')}</Button
 		>
 
 		{#if platform === 'web'}
-			<Button on:click={saveContent} variant="outline" class="mt-6 p-5"
+			<Button onclick={saveContent} variant="outline" class="mt-6 p-5"
 				>{$t('settings.save-to-webstore')}</Button
 			>
 		{/if}
@@ -116,40 +128,38 @@
 		<Label for="select" class="text-lg">{$t('settings.general')}</Label>
 		<div class="grid gap-2 px-2">
 			<Select.Root
-				selected={{
-					label: $t('settings.languages.' + $locale?.split('-')[0]),
-					value: $locale
-				}}
-				onSelectedChange={onLanguageSelectedChanged}
+				type="single"
+				bind:value={selectedLanguage as string}
+				onValueChange={onLanguageSelectedChanged}
 			>
 				<div class="flex flex-row items-center">
 					<p class="flex-1">{$t('settings.language')}</p>
-					<Select.Trigger class="w-1/2" id="select">
-						<Select.Value />
-					</Select.Trigger>
+					<Select.Trigger class="w-1/2" id="select">{languageTriggerContent}</Select.Trigger>
 				</div>
 				<Select.Content>
 					{#each $locales as l}
-						<Select.Item value={l}>{$t('settings.languages.' + l)}</Select.Item>
+						<Select.Item value={l} label={$t('settings.languages.' + l)}
+							>{$t('settings.languages.' + l)}</Select.Item
+						>
 					{/each}
 				</Select.Content>
 			</Select.Root>
 			<Select.Root
-				selected={{
-					label: $t('settings.themes.' + $globalTheme),
-					value: $globalTheme
-				}}
-				onSelectedChange={onThemeSelected}
+				type="single"
+				bind:value={selectedTheme as string}
+				onValueChange={onThemeSelected}
 			>
 				<div class="flex flex-row items-center">
 					<p class="flex-1">{$t('settings.theme-label')}</p>
 					<Select.Trigger class="w-1/2" id="select">
-						<Select.Value />
+						{themeTriggerContent}
 					</Select.Trigger>
 				</div>
 				<Select.Content>
 					{#each themeSettings as theme}
-						<Select.Item value={theme}>{$t('settings.themes.' + theme)}</Select.Item>
+						<Select.Item value={theme} label={$t('settings.themes.' + theme)}
+							>{$t('settings.themes.' + theme)}</Select.Item
+						>
 					{/each}
 				</Select.Content>
 			</Select.Root>
@@ -172,18 +182,16 @@
 					<div class="flex flex-row items-center">
 						<p class="flex-1">Interval</p>
 						<Select.Root
-							selected={{
-								label: intervalTimes.get($notificationInterval),
-								value: $notificationInterval
-							}}
-							onSelectedChange={onIntervalChanged}
+							type="single"
+							bind:value={selectedInterval}
+							onValueChange={onIntervalChanged}
 						>
 							<Select.Trigger class="w-48">
-								<Select.Value />
+								{intervalTriggerContent}
 							</Select.Trigger>
 							<Select.Content>
 								{#each intervalTimes as [key, value]}
-									<Select.Item value={key}>{value}</Select.Item>
+									<Select.Item value={key} label={value}>{value}</Select.Item>
 								{/each}
 							</Select.Content>
 						</Select.Root>
