@@ -1,6 +1,6 @@
 <script lang="ts">
 	import SubjectCard from './SubjectCard.svelte';
-	import Button from '$lib/components/ui/button/button.svelte';
+	import Button, { buttonVariants } from '$lib/components/ui/button/button.svelte';
 	import { dataService } from '$lib/database';
 	import * as Drawer from '$lib/components/ui/drawer';
 	import * as Popover from '$lib/components/ui/popover';
@@ -14,10 +14,13 @@
 	import { currentTab, subjects } from '$lib/stores';
 	import type { Subject } from '$lib/Subject';
 	import { toast } from 'svelte-sonner';
+	import { clsx } from 'clsx';
 
-	$: $currentTab = $t('titles.subjects');
+	$effect(() => {
+		$currentTab = $t('titles.subjects');
+	});
 
-	let colors = [
+	let colors = $state<string[]>([
 		'#00E015',
 		'#E0A100',
 		'#003BE0',
@@ -34,21 +37,22 @@
 		'#613045',
 		'#093949',
 		'#f08561'
-	];
+	]);
 
-	let selectedColor = 0;
-	let selectedName = 'Mathe';
-	let drawerOpen = false;
-	let editingSubject = false;
+	let selectedColor = $state(0);
+	let selectedName = $state('Mathe');
+	let drawerOpen = $state(false);
+	let editingSubject = $state(false);
+
 	let currentSubject: Subject | null = null;
 
-	function onChangeColor(event: { detail: { color: string } }) {
-		colors[colors.length - 1] = event.detail.color;
+	function onChangeColor(color: string) {
+		colors[colors.length - 1] = color;
 	}
 
 	function addColor() {
-		colors = [...colors, '#ffffff'];
 		selectedColor = colors.length - 1;
+		colors.push('#ffffff');
 	}
 
 	async function submittedNewSubject() {
@@ -113,7 +117,8 @@
 		drawerOpen = true;
 	}
 
-	function onSubmit() {
+	function onSubmit(e: Event) {
+		e.preventDefault();
 		if (editingSubject) {
 			editSubject();
 		} else {
@@ -128,10 +133,10 @@
 			<h3 class="scroll-m-20 text-2xl font-semibold tracking-tight">
 				{$t('subjects.no-subjects')}
 			</h3>
-			<Drawer.Trigger asChild let:builder>
-				<Button builders={[builder]} variant="outline" class="h-16 transition active:scale-95"
-					><Plus class="mr-3" />{$t('subjects.new-subject')}</Button
-				>
+			<Drawer.Trigger
+				class={clsx(buttonVariants({ variant: 'outline' }), 'h-16 transition active:scale-95')}
+			>
+				<Plus class="mr-3" />{$t('subjects.new-subject')}
 			</Drawer.Trigger>
 		</div>
 	{:else}
@@ -141,17 +146,16 @@
 					id={subject.id}
 					name={subject.name}
 					color={subject.color}
-					on:edit={onEditSubject}
+					onedit={onEditSubject}
 				/>
 			{/each}
-			<Drawer.Trigger asChild let:builder>
-				<Button
-					builders={[builder]}
-					variant="outline"
-					on:click={() => (editingSubject = false)}
-					class="h-16 w-full transition active:scale-95"
-					><Plus class="mr-3" />{$t('subjects.new-subject')}</Button
-				>
+			<Drawer.Trigger
+				onclick={() => (editingSubject = false)}
+				class={clsx(
+					buttonVariants({ variant: 'outline' }),
+					'h-16 w-full transition active:scale-95'
+				)}
+				><Plus class="mr-3" />{$t('subjects.new-subject')}
 			</Drawer.Trigger>
 		</div>
 	{/if}
@@ -164,8 +168,8 @@
 				<Drawer.Title>{$t('subjects.drawer.new-subject.title')}</Drawer.Title>
 			{/if}
 		</Drawer.Header>
-		<Popover.Root>
-			<form class="mb-1 grid items-start gap-4 px-4" on:submit|preventDefault={onSubmit}>
+		<Popover.Root onOpenChange={(open) => open && addColor()}>
+			<form class="mb-1 grid items-start gap-4 px-4" onsubmit={onSubmit}>
 				<div class="grid gap-2">
 					<Label for="name">{$t('subjects.drawer.name-label')}</Label>
 					<Input autocomplete="off" type="text" id="name" bind:value={selectedName} />
@@ -174,38 +178,39 @@
 				<div class="flex flex-wrap gap-2 p-2">
 					{#each colors as color, index (index)}
 						<button
+							aria-label={color}
 							style="background-color: {color};"
 							class="aspect-square w-10 rounded-full border-0 border-foreground"
 							class:selected={index === selectedColor}
 							type="button"
-							on:click={() => (selectedColor = index)}
+							onclick={() => (selectedColor = index)}
 						></button>
 					{/each}
-					<Popover.Trigger asChild let:builder>
-						<Button
-							on:click={addColor}
-							variant="outline"
-							builders={[builder]}
-							class="aspect-square rounded-full p-0"><Plus size={16} /></Button
-						>
+					<Popover.Trigger
+						class={clsx(
+							buttonVariants({ variant: 'outline', size: 'icon' }),
+							'aspect-square !rounded-full  p-0'
+						)}
+					>
+						<Plus size={16} />
 					</Popover.Trigger>
 				</div>
 			</form>
 			<Popover.Content>
-				<Colorpicker on:change={onChangeColor} />
+				<Colorpicker onValueChange={onChangeColor} />
 			</Popover.Content>
 		</Popover.Root>
 		<Drawer.Footer class="pt-2">
 			{#if editingSubject}
-				<Button on:click={editSubject}>{$t('ui.save')}</Button>
-				<Drawer.Close asChild let:builder>
-					<Button variant="outline" builders={[builder]}>{$t('ui.cancel')}</Button>
+				<Button onclick={editSubject}>{$t('ui.save')}</Button>
+				<Drawer.Close class={buttonVariants({ variant: 'outline' })}>
+					{$t('ui.cancel')}
 				</Drawer.Close>
-				<Button variant="destructive" on:click={removeSubject}>{$t('ui.delete')}</Button>
+				<Button variant="destructive" onclick={removeSubject}>{$t('ui.delete')}</Button>
 			{:else}
-				<Button on:click={submittedNewSubject}>{$t('ui.save')}</Button>
-				<Drawer.Close asChild let:builder>
-					<Button variant="outline" builders={[builder]}>{$t('ui.cancel')}</Button>
+				<Button onclick={submittedNewSubject}>{$t('ui.save')}</Button>
+				<Drawer.Close class={buttonVariants({ variant: 'outline' })}>
+					{$t('ui.cancel')}
 				</Drawer.Close>
 			{/if}
 		</Drawer.Footer>
