@@ -1,6 +1,6 @@
-import Root from './checkbox.svelte';
+import Root from "./checkbox.svelte";
 export {
 	Root,
 	//
-	Root as Checkbox
+	Root as Checkbox,
 };
