@@ -1,17 +1,17 @@
-import { Drawer as DrawerPrimitive } from 'vaul-svelte';
+import { Drawer as DrawerPrimitive } from "vaul-svelte";
 
-import Root from './drawer.svelte';
-import Content from './drawer-content.svelte';
-import Description from './drawer-description.svelte';
-import Overlay from './drawer-overlay.svelte';
-import Footer from './drawer-footer.svelte';
-import Header from './drawer-header.svelte';
-import Title from './drawer-title.svelte';
-import NestedRoot from './drawer-nested.svelte';
+import Root from "./drawer.svelte";
+import Content from "./drawer-content.svelte";
+import Description from "./drawer-description.svelte";
+import Overlay from "./drawer-overlay.svelte";
+import Footer from "./drawer-footer.svelte";
+import Header from "./drawer-header.svelte";
+import Title from "./drawer-title.svelte";
+import NestedRoot from "./drawer-nested.svelte";
 
-const Trigger = DrawerPrimitive.Trigger;
-const Portal = DrawerPrimitive.Portal;
-const Close = DrawerPrimitive.Close;
+const Trigger: typeof DrawerPrimitive.Trigger = DrawerPrimitive.Trigger;
+const Portal: typeof DrawerPrimitive.Portal = DrawerPrimitive.Portal;
+const Close: typeof DrawerPrimitive.Close = DrawerPrimitive.Close;
 
 export {
 	Root,
@@ -37,5 +37,5 @@ export {
 	Title as DrawerTitle,
 	Trigger as DrawerTrigger,
 	Portal as DrawerPortal,
-	Close as DrawerClose
+	Close as DrawerClose,
 };
