@@ -1,5 +1,5 @@
-import { Popover as PopoverPrimitive } from 'bits-ui';
-import Content from './popover-content.svelte';
+import { Popover as PopoverPrimitive } from "bits-ui";
+import Content from "./popover-content.svelte";
 const Root = PopoverPrimitive.Root;
 const Trigger = PopoverPrimitive.Trigger;
 const Close = PopoverPrimitive.Close;
@@ -13,5 +13,5 @@ export {
 	Root as Popover,
 	Content as PopoverContent,
 	Trigger as PopoverTrigger,
-	Close as PopoverClose
+	Close as PopoverClose,
 };
