@@ -1,7 +1,7 @@
-import { Tabs as TabsPrimitive } from 'bits-ui';
-import Content from './tabs-content.svelte';
-import List from './tabs-list.svelte';
-import Trigger from './tabs-trigger.svelte';
+import { Tabs as TabsPrimitive } from "bits-ui";
+import Content from "./tabs-content.svelte";
+import List from "./tabs-list.svelte";
+import Trigger from "./tabs-trigger.svelte";
 
 const Root = TabsPrimitive.Root;
 
@@ -14,5 +14,5 @@ export {
 	Root as Tabs,
 	Content as TabsContent,
 	List as TabsList,
-	Trigger as TabsTrigger
+	Trigger as TabsTrigger,
 };
