@@ -1,19 +1,19 @@
 <script lang="ts">
-	import { Tabs as TabsPrimitive } from 'bits-ui';
-	import { cn } from '$lib/utils.js';
+	import { Tabs as TabsPrimitive } from "bits-ui";
+	import { cn } from "$lib/utils.js";
 
-	type $$Props = TabsPrimitive.ListProps;
-
-	let className: $$Props['class'] = undefined;
-	export { className as class };
+	let {
+		ref = $bindable(null),
+		class: className,
+		...restProps
+	}: TabsPrimitive.ListProps = $props();
 </script>
 
 <TabsPrimitive.List
+	bind:ref
 	class={cn(
-		'inline-flex h-10 items-center justify-center rounded-md bg-muted p-1 text-muted-foreground',
+		"bg-muted text-muted-foreground inline-flex h-10 items-center justify-center rounded-md p-1",
 		className
 	)}
-	{...$$restProps}
->
-	<slot />
-</TabsPrimitive.List>
+	{...restProps}
+/>
